requirements.txt
accelerate==1.6.0
diffusers==0.33.1
transformers==4.46.2
gradio==5.23.0
sentencepiece==0.2.0
pillow==11.1.0
av==12.1.0
numpy==1.26.2
scipy==1.12.0
requests==2.31.0
torchsde==0.2.6

einops
opencv-contrib-python
safetensors

djangorestframework>=3.14.0
django-cors-headers>=4.2.0
celery[redis]>=5.3.6
redis>=5.0.3
torch>=2.3.0
boto3
django-storages
torchvision